LFS-818: Display an estimate of the total number of results in the livetable
Add the isApprox field to the JSON returned from the PaginationServlet and set it true if totalrows is a lower-bound estimate of the total number of query matches. Set it to false if totalrows is an exact value of the number of query matches.

Use the isApprox field as the means to signal that totalrows is an approximate value instead of returning totalrows as a negative value.