LFS-859: Make the color of sidebar menu items configurable
The color is specified similarly to the app name and version, and is accessible to the frontend code via a meta tag.
Also added support for a more versatile shade of blue to replace the default teal, and for a shade a pink.
The colors for each project were set as follows:
* LFS - blue (new shade)
* Cards4Care - orange
* Cards4kids - purple