CARDS-2022: Add support for a Docker Compose environment with a reverse proxy that listens on two ports, one for administrative operations, one for user operations
Create the cloud-iam-demo-saml-ssl OSGi module which is the same as cloud-iam-demo-saml except that it is for a CARDS instance available at https://localhost as opposed to http://localhost:8080. If generate_compose_yaml.py is called with --saml_cloud_iam_demo and the --ssl_proxy flag is specified, use cloud-iam-demo-saml-ssl instead of cloud-iam-demo-saml.

This is needed for testing this PR