CARDS-2221: Create a script within the CARDS repository for building a self-contained Docker image using the cards-sling-feature-downloader utility
"mvn clean install -Pdocker" will create a Docker image named "cards/cards:latest" if this repository is checked out to a version of CARDS that ends in "-SNAPSHOT". Otherwise, the version of CARDS that this repository is checked out to (eg. "0.9.7", "0.9.8", etc...) will be used for naming the CARDS Docker image (eg. "cards/cards:0.9.7", "cards/cards:0.9.8", etc...)