CARDS-59: Add support for running frontend tests using Selenium Add selenium test module. This module can be run using `mvn clean install -Dselenium`. It runs tests using Docker to launch CARDS and Google Chrome to run the tests. To support this, `-Dselenium` will add docker to the build in addition to cleaning out any running or old CARDS docker builds.
To run these tests, make sure that you have Google Chrome installed
in the default install location for your operating system. The
expected location can be found here:
https://github.com/SeleniumHQ/selenium/wiki/ChromeDriver/01fde32d0ed245141e24151f83b7c2db31d596a4#requirements

If you are using WSL, make sure that Google Chrome is installed
through WSL, not just through Windows:
https://scottspence.com/posts/use-chrome-in-ubuntu-wsl

The current test is to login and verify that the user's avatar on
the top right of the dashboard is correct. To ensure that this test
ran, you can check `/selenium/target/failsafe-reports`. There should
be 1 run test and no failures.

If you want to check behaviour if there is a failed test, modify
`LoginSelenium.java` and change `assertEquals("A"` to some other
string.